pyproject.toml

[project]
name = "weiser-ai"
version = "0.1.15"
description = "Enterprise-grade data quality framework with YAML configuration, LLM-friendly design, and advanced statistical validation"
keywords = ["data-quality", "data-validation", "data-testing", "sql", "yaml", "llm", "ai"]
authors = [
    {name = "Paco Valdez", email = "paco.valdez@berkeley.edu"},
]
maintainers = [
    {name = "Paco Valdez", email = "paco.valdez@berkeley.edu"},
]
dependencies = [
    "typer[all]>=0.9.0",
    "sqlglotrs==0.1.0",
    "pydantic>=2.5.3",
    "sqlalchemy>=2.0.41",
    "duckdb>=0.9.2",
    "jinja2>=3.1.2",
    "python-dotenv>=1.0.0",
    "psycopg2>=2.9.5",
    "pandas>=2.2.2",
    "boto3>=1.35.49",
    "sqlglot==20.5.0",
    "pyyaml==6.0.2",
    "slack-sdk>=3.34.0",
    "snowflake-sqlalchemy>=1.7.5",
    "databricks-sqlalchemy>=2.0.7",
    "sqlalchemy-bigquery>=1.13.0",
]
requires-python = "<3.13,>=3.10"
readme = "README.md"
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]

[project.urls]
"Homepage" = "https://weiser.ai"
"Documentation" = "https://weiser.ai"
"Repository" = "https://github.com/weiser-ai/weiser-ai"
"Bug Tracker" = "https://github.com/weiser-ai/weiser-ai/issues"

[project.scripts]
weiser = "weiser.main:app"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
source-includes = ["weiser/"]

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.4.0",
]
